% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildcompnet.R
\name{buildcompnet}
\alias{buildcompnet}
\title{Network models of interspecific competition with presence-absence data}
\usage{
buildcompnet(
  presabs,
  spvars_no_int = NULL,
  spvars_dist_int = NULL,
  spvars_multi_int = NULL,
  spvars_cat_no_int = NULL,
  spvars_cat_int = NULL,
  pairvars = NULL,
  rank = 0,
  family = "binomial",
  olre = TRUE,
  prior_intercept_scale = 5,
  prior_betas_scale = 5,
  prior_sigma_addeff_rate = 1,
  prior_sigma_olre_rate = 1,
  prior_lambda_scale = 5,
  warmup = 1000,
  iter = 2000,
  adapt_delta = 0.8
)
}
\arguments{
\item{presabs}{Must be specified by user. Binary (0 or 1) presence-absence matrix with sites as
rows and species as columns. Column names should be unique species names.}

\item{spvars_no_int}{A matrix or data frame, in which rows are species and columns are traits to
be included in the model as additive species effects only, with no interaction term. Row names
should be unique species names, and column names should be unique trait names.}

\item{spvars_dist_int}{A matrix or data frame, in which rows are species and columns are traits
to be included in the model with a distance interaction term (i.e., the absolute value of the
difference in trait values for each species pair). Row names should be unique species names,
and column names should be unique trait names.}

\item{spvars_multi_int}{A matrix or data frame, in which rows are species and columns are traits
to be included in the model with a multiplicative interaction term (i.e., the product of the
trait values for each species pair). Row names should be unique species names, and column
names should be unique trait names.}

\item{spvars_cat_no_int}{A matrix or data frame, in which rows are species and columns are categorical traits to
be included in the model as additive species effects only, with no interaction term. Row names
should be unique species names, and column names should be unique trait names.}

\item{spvars_cat_int}{A matrix or data frame, in which rows are species and columns are categorical traits to
be included in the model with a binary "same or different" interaction term. Row names
should be unique species names, and column names should be unique trait names.}

\item{pairvars}{A matrix or data frame, in which rows are species pairs and columns are
pair-level traits that do not have a species-level analog, e.g., phylogenetic distance.
Column names should be unique trait names. There
should also be two columns named "spAid" and "spBid" containing the unique names of the
species in each pair.}

\item{rank}{Number of dimensions for the multiplicative latent factor term. Rank=0 (the default)
yields a model with no multiplicative term.}

\item{family}{Distribution family for regression model. Defaults to 'binomial'. If family='fnchypg', or Fisher's noncentral
hypergeometric distribution. In this case, the quantity we are modeling, as a function of species
traits, dyadic random effects, etc. is Mainali et al.'s (2022, Science Advances) "alpha" or "co-occurrence affinity"
parameter in the fnchypg distribution. Link function is identity. If family='binomial', we are modeling p, the probability that both species co-occur at a given
site, given that at least one is present. Link function is logit.}

\item{olre}{Logical variable indicating whether to include observation-level random intercepts drawn from a
univariate normal distribution. Only relevant when family = 'fnchypg'. Defaults to TRUE, as these models
can often exhibit fit problems which may be mitigated by the random effects.}

\item{prior_intercept_scale}{Scale parameter for mean-zero Gaussian prior on the intercept term
for the linear predictor.}

\item{prior_betas_scale}{Scale parameter for mean-zero Gaussian priors on the coefficients of
fixed effect terms in the linear predictor.}

\item{prior_sigma_addeff_rate}{Rate parameter for exponential prior on the scale of the
species-level Gaussian random effects (i.e., "row and column effects").}

\item{prior_sigma_olre_rate}{Rate parameter for exponential prior on the scale of the
dyad-level Gaussian random effects.}

\item{prior_lambda_scale}{Scale parameter for mean-zero Gaussian prior on diagonal values of
Lambda, the matrix that determines how different species' values of the latent factors
interact in the linear predictor.}

\item{warmup}{Number of warmup iterations for Stan.}

\item{iter}{Number of posterior sampling iterations for Stan.}

\item{adapt_delta}{A parameter that tunes Stan's posterior sampling algorithm. Increasing closer to 1 can help avoid divergent transitions.}
}
\value{
Object of class "compnet", which is a list containing the stanfit model object,
a named list of posterior samples for all model parameters, a data frame containing all
input variables for the model, a matrix of dyadic X variables, a matrix of X variables
pertaining to species A in each pair, a matrix of X variables pertaining to species B in
each pair, and--when relevant--a matrix
of means and standard deviations for the input trait data before centering and scaling.
}
\description{
Network models of interspecific competition with presence-absence data
}
\details{
This function uses Stan, as accessed through the rstan package, to build a network
regression model of interspecific competitive niche differentiation in a Bayesian framework.
This function is designed to test the hypothesis that species are more likely to co-occur with
other species that are functionally dissimilar. Functional dissimilarity can be represented
directly by traits or by a proxy like phylogenetic distance. The core input data are a
species-by-site presence-absence matrix and one or more species-level traits
(e.g., plant leaf size) or pair-level traits (e.g., phylogenetic distance). Units of analysis
are species pairs. The response variable follows a binomial
distribution. The number of trials is the number of sites occupied by at least one
species in a pair, and the number of successes is the number of sites occupied by both species.
If species-level traits are used, each trait can be non-interacting
(i.e., there is no interaction term between species A's trait value and species B's),
interacting via a typical multiplicative term, or interacting via an absolute value difference
(i.e., "distance") term. The interaction terms are key to the core hypothesis. If competitive
niche differentiation is occurring, then the probability of co-occurrence is expected to
increase with trait or phylogenetic distance between species A and B, or with the product of
their trait values, if a multiplicative interaction is specified instead of a distance interaction.
Random effects are used to account for additive species-level dependencies and, optionally,
higher-order dependencies involving multiple species (e.g., "the enemy of my enemy is my friend").
Higher-order dependencies are modeled using a number of latent variable specified by "rank".
For more details on the random effects, see Hoff, P. (2021) Additive and multiplicative effects
network models. Stat. Sci. 36, 34–50.
}
\examples{

data(ex_presabs)
data(ex_traits)

# Quick demo run. Will prompt warnings.
# Run with default warmup and iter for good posterior sampling.
ex_compnet <- buildcompnet(presabs=ex_presabs,
spvars_dist_int=ex_traits[c("ndtrait")], warmup=10, iter=20)

}
