% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gofstats.R
\name{compute_gofstats}
\alias{compute_gofstats}
\title{Compute summary statistics quantifying row/column-level and third order dependencies in a
symmetric matrix with NA diagonals}
\usage{
compute_gofstats(Y)
}
\arguments{
\item{Y}{A symmetric matrix with NA diagonals}
}
\value{
A named vector containing: 1- the standard deviation of row means, and 2- The triadic
dependency metric used by Hoff, Fosdick, & Volfovsky's "amen" package.
}
\description{
Compute summary statistics quantifying row/column-level and third order dependencies in a
symmetric matrix with NA diagonals
}
\details{
Internal helper function.
}
