% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_phylo.R
\docType{data}
\name{ex_phylo}
\alias{ex_phylo}
\title{Phylogenetic distance between species pairs in a simulated metacommunity dataset with 24 species}
\format{
An object of class \code{data.frame} with 276 rows and 3 columns.
}
\usage{
ex_phylo
}
\description{
A data frame with 276 rows and 3 columns:
\describe{
\item{spAid}{ID for species A in each pair}
\item{spBid}{ID for species B in each pair}
\item{phylodist}{Phylogenetic distance between each species pair}
}
}
\keyword{datasets}
