% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_presabs.R
\docType{data}
\name{ex_presabs}
\alias{ex_presabs}
\title{A simulated metacommunity presence-absence matrix in which rows represent sites and columns represent species}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 100 rows and 24 columns.
}
\usage{
ex_presabs
}
\description{
Matrix with 100 rows and 24 columns containing 0 (absence) or 1 (presence)
}
\keyword{datasets}
