% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit.R
\name{expit}
\alias{expit}
\title{Inverse logit function}
\usage{
expit(x)
}
\arguments{
\item{x}{A numeric value, or a vector or array of numeric values.}
}
\value{
The inverse-logit(s) of the supplied numeric value(s).
}
\description{
Inverse logit function
}
\examples{

expit(-1)

exvec <- c(-2:2)
expit(exvec)

exmat <- matrix(c(1:12), nrow=3)
expit(exmat)


}
