% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpredsamp.R
\name{postpredsamp}
\alias{postpredsamp}
\title{Simulate posterior predictive samples from a compnet model}
\usage{
postpredsamp(mod)
}
\arguments{
\item{mod}{Object of class "compnet", which is created by the compnet() function.}
}
\value{
A matrix of posterior predictive samples with a row for each observation and a column for each sample.
}
\description{
Simulate posterior predictive samples from a compnet model
}
\examples{

data(ex_presabs)
data(ex_traits)

# Quick demo run. Will prompt warnings.
# Run with default warmup and iter for good posterior sampling.
ex_compnet <- compnet(presabs=ex_presabs, spvars_dist_int=ex_traits, warmup=10, iter=20)
ex_compnet_pps <- postpredsamp(ex_compnet)

}
