% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_pairvar.R
\name{scatter_pairvar}
\alias{scatter_pairvar}
\title{Deprecated. Please use plot_pairvar().}
\usage{
scatter_pairvar(
  mod,
  xvar,
  xlabel,
  color = "red",
  orig.scale = TRUE,
  ymin = 0,
  ymax = 1,
  ci_width = 0.95,
  grid_size = 100,
  thin = TRUE,
  thin_to = 100
)
}
\arguments{
\item{mod}{An object of class "compnet" created by the buildcompnet() function.}

\item{xvar}{Character string for the name of the trait to be used. Must match the trait name in
the input data used to build the model.}

\item{xlabel}{Optional character string to replace xvar when plotting.}

\item{color}{Color to use in plotting.}

\item{orig.scale}{Logical value indicating whether to back-transform trait data to the original
scale (TRUE) or leave them with mean zero and unit variance (FALSE).}

\item{ymin}{Real number indicating the location of the bottom of the plot's y axis.}

\item{ymax}{Real number indicating the location of the top of the plot's y axis.}

\item{ci_width}{A real number (0,1) describing the desired widths of credible band. Defaults to 0.95.}

\item{grid_size}{A positive integer defining the number of discrete steps to use in approximating
the shape of mean prediction curves and credible bands. Defaults to 100.}

\item{thin}{Logical value determining whether to use a random subsample of the full posterior sample.}

\item{thin_to}{Integer value determining how many random samples to draw from the full posterior sample if thin=TRUE.}
}
\description{
Deprecated. Please use plot_pairvar().
}
