% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_pairvar_getdata.R
\name{scatter_pairvar_getdata}
\alias{scatter_pairvar_getdata}
\title{Get data for a plot of the effect of a pair-level variable like phylogenetic distance}
\usage{
scatter_pairvar_getdata(
  mod,
  xvar,
  orig.scale = TRUE,
  ci_width = 0.95,
  grid_size = 100,
  thin = TRUE,
  thin_to = 100
)
}
\arguments{
\item{mod}{An object of class "compnet" created by the buildcompnet() function.}

\item{xvar}{Character string for the name of the trait to be used. Must match the trait name in
the input data used to build the model.}

\item{orig.scale}{Logical value indicating whether to back-transform trait data to the original
scale (TRUE) or leave them with mean zero and unit variance (FALSE).}

\item{ci_width}{A real number (0,1) describing the desired widths of credible band. Defaults to 0.95.}

\item{grid_size}{A positive integer defining the number of discrete steps to use in approximating
the shape of mean prediction curves and credible bands. Defaults to 100.}

\item{thin}{Logical value determining whether to use a random subsample of the full posterior sample.}

\item{thin_to}{Integer value determining how many random samples to draw from the full posterior sample if thin=TRUE.}
}
\value{
A ggplot2 graphic.
}
\description{
Get data for a plot of the effect of a pair-level variable like phylogenetic distance
}
\examples{
data(ex_presabs)
data(ex_phylo)

# Quick demo run. Will prompt warnings.
# Run with default warmup and iter for good posterior sampling.
ex_compnet_phylo <- buildcompnet(presabs=ex_presabs, pairvars=ex_phylo, warmup=10, iter=20)

scatter_pairvar_getdata(ex_compnet_phylo, xvar="phylodist")
}
