% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_compnet.R
\name{summarize_compnet}
\alias{summarize_compnet}
\title{Summarizing compnet model output}
\usage{
summarize_compnet(mod, ci_width = 0.95)
}
\arguments{
\item{mod}{Object of class "compnet", which is created by the compnet function.}
}
\value{
A data frame summarizing means and credible intervals for standardized effect sizes of fixed effects.
}
\description{
Summarizing compnet model output
}
\examples{

data(ex_presabs)
data(ex_traits)

# Quick demo run. Will prompt warnings.
# Run with default warmup and iter for good posterior sampling.
ex_compnet <- compnet(presabs=ex_presabs, spvars_dist_int=ex_traits, warmup=100, iter=200)

ex_compnet_summ <- summarize_compnet(ex_compnet)
ex_compnet_summ

}
