% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gofstats.R
\name{compute_gofstats}
\alias{compute_gofstats}
\title{Compute summary statistics quantifying row/column-level and third order dependencies in a
symmetric matrix with NA diagonals}
\usage{
compute_gofstats(Y)
}
\arguments{
\item{Y}{A symmetric matrix with NA diagonals}
}
\value{
A named vector containing: 1- the standard deviation of row means, and 2- The triadic
dependency metric used by Hoff, Fosdick, & Volfovsky's "amen" package.
}
\description{
Compute summary statistics quantifying row/column-level and third order dependencies in a
symmetric matrix with NA diagonals
}
\examples{

data(ex_presabs)
data(ex_traits)

# Quick demo run. Will prompt warnings.
# Run with default warmup and iter for good posterior sampling.
ex_compnet <- compnet(presabs=ex_presabs, spvars_dist_int=ex_traits, warmup=10, iter=20)

ex_compnet_summ <- summarize_compnet(ex_compnet)
ex_compnet_summ

}
