% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_traits.R
\docType{data}
\name{ex_traits}
\alias{ex_traits}
\title{Species-level data for two traits in a simulated metacommunity dataset with 24 species}
\format{
An object of class \code{data.frame} with 24 rows and 2 columns.
}
\usage{
ex_traits
}
\description{
A data frame with 24 rows and 2 columns:
\describe{
\item{ndtrait}{A trait driving competitive niche differentiation}
\item{domtrait}{A trait influencing competitive ability}
}
}
\keyword{datasets}
