% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofstats.R
\name{gofstats}
\alias{gofstats}
\title{Report quantiles of gofstats values for observed data relative to posterior predictive distribution}
\usage{
gofstats(mod, thin = T, thin_to = 300)
}
\arguments{
\item{mod}{An object of class "compnet" created by the buildcompnet() function.}

\item{thin}{Logical value indicating whether to take a random subsample of posterior draws.}

\item{thin_to}{Logical value indicating the size of the subsample to take if thin=TRUE.}
}
\value{
A named vector containing quantiles on the interval \eqn{[0,1]} for: 1- the standard deviation
of row means, and 2- The triadic dependency metric used by Hoff, Fosdick, & Volfovsky's "amen"
package. These values represent the proportion of the posterior predictive simulation that were
less than the value for the observed data.
}
\description{
Report quantiles of gofstats values for observed data relative to posterior predictive distribution
}
\details{
This function can be used to assess whether species-level and higher-order dependencies
in the data are represented adequately by the model structure. Extreme output values indicate
there may be a problem. In these cases, it may help to use different fixed effect predictors,
or to increase the model rank.
}
\examples{

data(ex_presabs)
data(ex_traits)

# Quick demo run. Will prompt warnings.
# Run with default warmup and iter for good posterior sampling.
ex_compnet <- buildcompnet(presabs=ex_presabs, spvars_dist_int=ex_traits[c("ndtrait")], warmup=10, iter=20)

gofstats(ex_compnet)

}
