% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbetabinom.R
\name{rbetabinom}
\alias{rbetabinom}
\title{Simulate draws from a beta-binomial distribution}
\usage{
rbetabinom(n, size, prob, phi)
}
\arguments{
\item{n}{Integer value: number of samples to draw.}

\item{size}{Integer value: number of trials for each sample.}

\item{prob}{Vector, matrix, or array of real values specifying the probability of success.}

\item{phi}{Real value: "over/under- dispersion parameter" for beta-binomial likelihood. In this
parameterization, the latent beta variable has shape \eqn{\mu\phi} and scale
\eqn{(1-\mu)\phi}, where \eqn{\mu} is the probability of success in the binomial distribution.}
}
\value{
A vector of draws from the specified beta-binomial distribution.
}
\description{
Simulate draws from a beta-binomial distribution
}
\examples{

data(ex_presabs)
data(ex_traits)

# Quick demo run. Will prompt warnings.
# Run with default warmup and iter for good posterior sampling.
ex_compnet <- compnet(presabs=ex_presabs, spvars_dist_int=ex_traits, warmup=10, iter=20)
ex_compnet_pps <- postpredsamp(ex_compnet)

}
